Since all phones have the same functionality, I modified the Phone interface to abstract class, where I implemented the methods. So when we deal with other manufacturers, we don't have to reimplement all functionality, just the color, material and IMEI would change.
