+++ src/Samsung.java
@@ -1,54 +1,21 @@
 
-public abstract class Samsung implements Phone {
+public abstract class Samsung extends Phone {
 
-    private final Integer batteryLife;
+    private String operatingSystem;
 
-    private String color;
-
-    private String material;
-
-    private String messageContent;
-
-    public Samsung(Integer batteryLife, String color, String material) {
-        this.batteryLife = batteryLife;
-        this.color = color;
-        this.material = material;
+    public Samsung(Integer batteryLife, String operatingSystem) {
+        super(batteryLife);
+        this.operatingSystem = operatingSystem;
     }
 
-    public String getMaterial() {
-        return material;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public Integer getBatteryLife() {
-        return batteryLife;
+    public String getOperatingSystem() {
+        return operatingSystem;
     }
 
     @Override
-    public void sendMessage(String phoneNumber, String messageContent) {
-    }
-
-    @Override
-    public void call(String phoneNumber) {
-
-    }
-
-    @Override
-    public void viewCallHistory() {
-
-    }
-
-    public abstract void addContact(Integer contactsOrder, String phoneNumber, String lastName, String firstName);
-
-    @Override
-    public void seeAllMessages() {
-
-    }
-
-    public void seeExistingContacts() {
-
+    public String toString() {
+        return "Samsung{" +
+                "operatingSystem='" + operatingSystem + '\'' +
+                '}';
     }
 }
